% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckUpdateFile.R
\name{check_if_update}
\alias{check_if_update}
\title{Check id file is recent}
\usage{
check_if_update(path_file, margin_days)
}
\arguments{
\item{path_file}{character variable}

\item{margin_days}{numeric variable}
}
\value{
string
}
\description{
Check if it has been updated for more than x days
}
