% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertDatePt.R
\name{convert_to_date_pt}
\alias{convert_to_date_pt}
\title{Convert Portuguese Date Strings to Date Objects}
\usage{
convert_to_date_pt(dates_vector, default_year = year(Sys.Date()))
}
\arguments{
\item{dates_vector}{A character vector containing the date strings to be converted.}

\item{default_year}{An integer specifying the default year to use when the date string does not include a year. Defaults to the current year.}
}
\value{
A `Date` vector where each element corresponds to the last valid date found in the respective string from `dates_vector`. If no valid date is found, the element will be `NA`.
}
\description{
This function processes a vector of date strings in Portuguese and converts them into `Date` objects.
It supports various date formats, Portuguese month names, and multiple dates within a single string.
When no year is specified in the date string, a default year can be applied.
}
\details{
- The function translates Portuguese month names into English for compatibility with R's date parsing functions.
- If a string contains multiple dates separated by delimiters (e.g., "e", "a", ";"), only the last date is returned.
- Non-parsable strings will result in `NA`.
- Common delimiters and structures like "até", "e", and commas are handled automatically.
}
\examples{
# Example date vector
dates_vector <- c("08 e 09 de Outubro", "09/04",
                  "26/07 a 31/07 e 06/08 a 10/08 de 2024",
                  "Text without dates", "15-10-2024 até 17-10-2024")

# Convert Portuguese date strings to Date objects
convert_to_date_pt(dates_vector)

}
