% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveAccent.R
\name{remove_accent}
\alias{remove_accent}
\title{Remove All Accents from Strings}
\usage{
remove_accent(str, pattern = "all")
}
\arguments{
\item{str}{A character vector to process.}

\item{pattern}{A character vector indicating which accent types to remove. Defaults to `"all"`, which removes all accents.}
}
\value{
A character vector with accents removed.
}
\description{
This function removes accents from strings, converting accented characters to their unaccented equivalents.
}
