% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveAccent.R
\name{remove_accent}
\alias{remove_accent}
\title{remove all accent from string}
\usage{
remove_accent(str, pattern = "all")
}
\arguments{
\item{str}{character variable}

\item{pattern}{numeric variable}
}
\value{
string
}
\description{
remove all accent from string
}
